+++ .npmignore
@@ -1,4 +1,6 @@
 .idea
 demo/html/build
 demo/node.js/build
+test/build
+test/generated
 node_modules
