+++ test/test.ts
@@ -1,29 +1,30 @@
 /// <reference path="typings/node/node.d.ts" />
-/// <reference path="typings/mocha/mocha-node.d.ts" />
 /// <reference path="typings/chai/chai.d.ts" />
 /// <reference path="../src/iq.ts" />
-
-declare var PNGImage : any;
-
-var chai : typeof chai = require('chai'),
-	path               = require("path"),
-	fs                 = require("fs");
-
-var PNGImage = require('pngjs-image');
-
-var expect = chai.expect;
+/// <reference path="typings/mocha/mocha.d.ts"/>
+/// <reference path="utils.ts"/>
 
 module Test {
+
+	var PNGImage : any     = require('pngjs-image'),
+		chai : typeof chai = require('chai'),
+		path               = require("path"),
+		fs                 = require("fs"),
+		expect             = chai.expect;
+
 	var options = {
 		colors   : [
-			64
+			2,
+			8,
+			16,
+			64,
+			256,
+			2048
 		],
 		image    : [
 			IQ.Image.NearestColor,
-			/*
-			 IQ.Image.ErrorDiffusionArray,
-			 IQ.Image.ErrorDiffusionRiemersma
-			 */
+			IQ.Image.ErrorDiffusionArray,
+			IQ.Image.ErrorDiffusionRiemersma
 		],
 		palette  : [
 			IQ.Palette.NeuQuant,
@@ -43,57 +44,6 @@ module Test {
 			IQ.Distance.ManhattanSRGB
 		]
 	};
-
-	function createDirectory(dir : string) {
-		var folders = path.normalize(dir).replace(/\\/g, "/").split("/");
-
-		if (folders && folders.length > 0) {
-			for (var i = 0; i < folders.length; i++) {
-				var testDir = folders.slice(0, i + 1).join("/");
-				if (!fs.existsSync(testDir)) {
-					fs.mkdirSync(testDir);
-				}
-			}
-		}
-	}
-
-	function getFilesInFolder(folder : string, recursive : boolean = true, filter : (file : string) => boolean = null, subFolder : string = null) : string[] {
-		var fullFolder  = subFolder === null ? folder : path.join(folder, subFolder),
-			folderFiles = fs.readdirSync(fullFolder),
-			files       = [];
-
-		folderFiles.forEach(function (file) {
-			if (filter && filter(file)) {
-				console.log(path.join(fullFolder, file) + " removed by filter");
-				return;
-			}
-
-			var stat              = fs.statSync(path.join(fullFolder, file)),
-				subFolderFileName = subFolder === null ? file : path.join(subFolder, file);
-
-			if (stat.isFile()) {
-				files.push(subFolderFileName);
-			} else if (stat.isDirectory()) {
-				if (recursive) {
-					files = files.concat(getFilesInFolder(folder, recursive, filter, subFolderFileName));
-				}
-			}
-		});
-
-		return files.map(function (file) {
-			return file.replace(/\\/g, "/");
-		});
-	}
-
-	function processLater(fn : () => any) {
-		setTimeout(fn, 1);
-	}
-
-	interface FileRecord {
-		file : string;
-		image : any;
-		pointContainer : IQ.Utils.PointContainer;
-	}
 
 	function testPalette(folder : string, fileRecords : FileRecord[], imageQuantizer : IQ.Image.IImageDitherer, paletteQuantizer : IQ.Palette.IPaletteQuantizer, distance : IQ.Distance.IDistanceCalculator, callback : () => void) : void {
 		folder = path.join("generated", folder);
@@ -138,61 +88,35 @@ module Test {
 		});
 	}
 
-	function readSourceFiles(sourceFolder : string, callback : (fileRecords) => void) {
-		var files       = getFilesInFolder("PngSuite", true, file => file.indexOf(".png") === -1),
-			fileRecords = [],
-			processed   = 0;
-
-		files.forEach(file => {
-			PNGImage.readImage(path.join(sourceFolder, file), function (err, image) {
-				if (err) {
-					console.log("ERROR - " + err + ", " + file);
-				} else {
-					var w   = image.getWidth(),
-						h   = image.getHeight(),
-						buf = new IQ.Utils.PointContainer(w, h);
-
-					for (var x = 0; x < w; x++) {
-						for (var y = 0; y < h; y++) {
-							var idx = image.getIndex(x, y);
-							buf.setAt(IQ.Utils.Point.createByRGBA(
-								image.getRed(idx),
-								image.getGreen(idx),
-								image.getBlue(idx),
-								image.getAlpha(idx)
-							), x, y);
-						}
-					}
-					console.log(file + " read OK");
-					fileRecords.push({
-						folder         : sourceFolder,
-						file           : file,
-						image          : image,
-						pointContainer : buf
-					});
-				}
-				processed++;
-				if (files.length === processed) {
-					processLater(() => {
-						callback(fileRecords);
-					});
-				}
-			});
-		});
-	}
-
-	function testConditionColors(fileRecords, colors : number, done : MochaDone) {
+	function testConditionColors(fileRecords, done : (err?) => void) {
 		var combinations = [];
 
-		options.palette.forEach(paletteKlass => {
-			options.image.forEach(imageKlass => {
-				options.distance.forEach(distanceKlass => {
-					fileRecords.forEach(fileRecord => {
-						combinations.push({
-							paletteKlass  : paletteKlass,
-							imageKlass    : imageKlass,
-							distanceKlass : distanceKlass,
-							fileRecord    : fileRecord
+		options.colors.forEach(colors => {
+			options.palette.forEach(paletteKlass => {
+				options.image.forEach(imageKlass => {
+					options.distance.forEach(distanceKlass => {
+						fileRecords.forEach(fileRecord => {
+							if (<any>imageKlass === IQ.Image.ErrorDiffusionArray) {
+								for (var i = 0; i <= 8; i++) {
+									combinations.push({
+										colors               : colors,
+										paletteKlass         : paletteKlass,
+										imageKlass           : imageKlass,
+										distanceKlass        : distanceKlass,
+										fileRecord           : fileRecord,
+										errorDiffusionKernel : i
+									});
+								}
+							} else {
+								combinations.push({
+									colors               : colors,
+									paletteKlass         : paletteKlass,
+									imageKlass           : imageKlass,
+									distanceKlass        : distanceKlass,
+									fileRecord           : fileRecord,
+									errorDiffusionKernel : -1
+								});
+							}
 						});
 					});
 				});
@@ -200,34 +124,26 @@ module Test {
 		});
 
 		var runNext = () => {
-			console.log("left ", combinations.length, " tasks!");
 			if (combinations.length > 0) {
 				var item          = combinations.pop(),
 					distanceKlass = item.distanceKlass,
 					imageKlass    = item.imageKlass,
 					paletteKlass  = item.paletteKlass,
-					fileRecord = item.fileRecord;
+					fileRecord    = item.fileRecord,
+					colors        = item.colors;
 
 				var distance = new distanceKlass(),
 					palette  = new (<any>paletteKlass)(distance, colors),
 					folder   = colors + "/" + (<any>imageKlass).name + "/" + (<any>paletteKlass).name + "/" + (<any>distanceKlass).name;
 
-				/*
-				 if (<any>imageKlass === IQ.Image.ErrorDiffusionArray) {
-				 for (var i = 0; i <= 8; i++) {
-				 var imageQuantizer = new (<any>imageKlass)(distance, i);
-				 n++;
-				 console.log("add");
-				 testPalette(folder + "/" + IQ.Image.ErrorDiffusionArrayKernel[i], fileRecords, imageQuantizer, palette, distance, () => {
-				 n--;
-				 checkDone();
-				 });
-				 }
-				 } else {
-				 */
-				var imageQuantizer = new (<any>imageKlass)(distance);
+				console.log(path.join(folder, fileRecord.file) + " (" + combinations.length + " tasks left)");
+				var imageQuantizer;
+				if (item.errorDiffusionKernel >= 0) {
+					imageQuantizer = new (<any>imageKlass)(distance, item.errorDiffusionKernel);
+				} else {
+					imageQuantizer = new (<any>imageKlass)(distance);
+				}
 				testPalette(folder, [fileRecord], imageQuantizer, palette, distance, runNext);
-				//}
 			} else {
 				done();
 			}
@@ -236,50 +152,13 @@ module Test {
 		runNext();
 	}
 
-	describe('image-quantization', () => {
-		var fileRecords;
-
-		before(done => {
-			readSourceFiles("PngSuite", _fileRecords => {
-				fileRecords = _fileRecords;
-				done();
-			});
+	readSourceFiles("images", fileRecords => {
+		testConditionColors(fileRecords, err => {
+			if (err) {
+				console.log(err);
+			} else {
+				console.log("done");
+			}
 		});
-
-		describe('color tests', function () {
-			options.colors.forEach(colors => {
-				it(colors + ' colors', function (done : MochaDone) {
-					this.timeout(50000);
-					testConditionColors(fileRecords, colors, done);
-				});
-			});
-		});
-		/*
-		 describe('image-quantization', function () {
-		 it('2 colors', function (done : MochaDone) {
-		 testConditionColors(fileRecords, 2, done);
-		 });
-		 });
-
-		 describe('image-quantization', function () {
-		 it('16 colors', function (done : MochaDone) {
-		 testConditionColors(fileRecords, 16, done);
-		 });
-		 });
-
-		 describe('image-quantization', function () {
-		 this.timeout(5000);
-		 it('64 colors', function (done : MochaDone) {
-		 testConditionColors(fileRecords, 64, done);
-		 });
-		 });
-
-		 describe('image-quantization', function () {
-		 this.timeout(50000);
-		 it('256 colors', function (done : MochaDone) {
-		 testConditionColors(fileRecords, 256, done);
-		 });
-		 });
-		 */
 	});
 }
